third_party/blink/web_tests/external/wpt/web-animations/responsive/perspective-expected.txt

This is a testharness.js-based test.
PASS perspective responsive to style changes
FAIL perspective clamped to 0px on keyframes assert_equals: expected "20px" but got "0px"
PASS perspective responsive to inherited changes
Harness: the test ran to completion.
